chore: extract getOptions to options.ts module
